only remove users if there are really no installations left (i.e. only on real uninstall)

+++ sia/minebox-sia.spec
@@ -43,8 +43,10 @@ set -e
 fi
 
 %postun
+if [ "$1" = 0 ] ; then
 /usr/sbin/userdel sia
 /usr/sbin/groupdel sia
+fi
 systemctl daemon-reload
 
 %files
